PLT-4790: Mention don't appear linked in message box preview mode. (#205)
* Added @here in getMentionKeys

* Moved postMessageHtmlToComponent post_message_view to post_utils for use in textbox

* Divided preview from return to render

* Update post_message_view.jsx
